[#2] Changed library build system to not make use of weak symbols.
Previously, C implementations of lib_nn kernels were defined with the same symbol name as the ASM implementations but were marked as weak symbols, causing the linker to prefer the ASM implementations if they were compiled. This is now handled explicitly through build flags.

In particular, if the library source is compiled with '-DNN_USE_REF', the C implementation will be used. Otherwise, the ASM is used.

Additionally, the make behavior now builds lib_nn as a static library (lib_nn.a) which can be built independently or built recursively via the unit tests (or other app) and linked into the application.