[#4] Fixes bugs in nn_conv2d_hstrip_tail_deep() and nn_conv2d_hstrip_tail_deep_padded().
The assembly implementations of those two functions misbehaved under certain conditions. In particular, the conditions seem to be that the input image has no input channel tail (channel count is multiple of 32), and there are multiple non-padding columns within the convolution window. It seems likely that there are additional conditions that would need to have been met, because otherwise I expect this would have manifested sooner.