Added AvgPool2d_Generic and AvgPool2d_Valid unit tests.
The tests for Generic that involve padding are currently failing. This is because of an incompatibility between our behavior and the behavior that TFLite uses when the pooling window intersects padding. TFLite just doesn't count the padding pixels and divides by a smaller number. We just use the padding value and divide by the same number of pixels as when there's no padding.